Tutorial Level Layout
Created the basic layout of the Tutorial level and set up a generic
tracking camera for the Tech Demo. There's a lot of individual code to
be cleaned up - especially when it comes to the protagonist. As it is,
this is just one room with the bare minimum assets.

Further details will be in the changelog
Ask me if you've got any questions!